pyproject.toml
[project]
name = "amac"
version = "0.2.1"
description = "AMAC — API Mapper + Auth Checker"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [{ name = "You" }]

dependencies = [
  "typer>=0.12,<0.13",
  "rich>=13.7,<14",
  "pydantic>=2.5,<3",
  "orjson>=3.9,<4",
  "httpx>=0.27,<0.28",
  "PyYAML>=6.0,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
  "ruff>=0.5,<0.6",
  "black>=24.3,<25",
]

[project.scripts]
amac = "amac.cli:app"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
ignore = ["E501"]
